+++ Policyfile.rb
@@ -10,9 +10,7 @@ name 'app_solr'
 default_source :supermarket
 
 # run_list: chef-client will run these recipes in the order specified.
-run_list 'app_solr::default'
-
-named_run_list :test, 'test::default'
+run_list 'test::default'
 
 # Specify a custom source for a single cookbook:
 cookbook 'app_solr', path: '.'