+++ resources/core.rb
@@ -2,7 +2,7 @@
 # Cookbook:: app_solr
 # Resource:: standalone_core
 #
-# Copyright:: 2021, Earth U
+# Copyright:: 2022, Earth U
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,10 +16,21 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+require 'pathname'
 cb = 'app_solr'
+
 unified_mode true
 
 # Create a core in standalone Solr
+
+# Check status of local standlone Solr core:
+#   curl 'http://localhost:8983/solr/admin/cores?action=STATUS&wt=json&core=corename'
+
+# By default, this resource creates a core using version 7.7.3
+# solrconfig.xml and schema.xml files. Custom files can be provided
+# as properties.
+# The cookbook also includes default solrconfig.xml and schema.xml files
+# for each version: 5.5.5, 6.6.6, and 7.7.3.
 
 property :name, String,
          description: 'Name of Solr core to create',
@@ -33,9 +44,16 @@ property :solr_user, String,
          description: 'Name of Solr user',
          default: 'solr'
 
+# Used only to provide a consistent :solr_home property
+# between the app_solr_core and app_solr_standalone resources.
+property :solr_dir, String,
+         description: 'Main directory for Solr. '\
+                      'Only used if :solr_home is relative.',
+         default: '/var/solr'
+
 property :solr_home, String,
-         description: 'Location of Solr home',
-         default: '/var/solr/data'
+         description: 'Location of Solr home. Can be relative to :solr_dir.',
+         default: 'data'
 
 property :use_custom_solrconfig, [true, false],
          description: 'Whether or not to use a custom solrconfig.xml. If '\
@@ -55,7 +73,8 @@ property :solrconfig_cookbook, String,
 property :use_custom_schema, [true, false],
          description: 'Whether or not to use a custom schema.xml. If '\
                       'true, provide values for :schema_source and '\
-                      ':schema_cookbook.',
+                      ':schema_cookbook. Useful if using the '\
+                      '"ClassicIndexSchemaFactory" schemaFactory in :solrconfig ',
          default: false
 
 property :schema_source, String,
@@ -68,20 +87,19 @@ property :schema_cookbook, String,
          default: cb
 
 action_class do
-  def solr_bin
-    "#{new_resource.extract_dir}/solr/bin/solr"
-  end
-
-  def core_dir
-    "#{new_resource.solr_home}/#{new_resource.name}"
-  end
-
-  def core_conf_dir
-    "#{core_dir}/conf"
+  def form_solr_path(x)
+    if x == ''
+      new_resource.solr_dir
+    else
+      Pathname.new(x).absolute? ? x : "#{new_resource.solr_dir}/#{x}"
+    end
   end
 end
 
 action :create do
+  core_dir = "#{form_solr_path(new_resource.solr_home)}/#{new_resource.name}"
+
+  solr_bin = "#{new_resource.extract_dir}/solr/bin/solr"
   execute "create_solr_core_#{new_resource.name}" do
     command "#{solr_bin} create -c #{new_resource.name}"
     user    new_resource.solr_user
@@ -89,7 +107,7 @@ action :create do
     not_if  { ::File.exist?("#{core_dir}/core.properties") }
   end
 
-  cookbook_file "#{core_conf_dir}/solrconfig.xml" do
+  cookbook_file "#{core_dir}/conf/solrconfig.xml" do
     source   new_resource.solrconfig_source
     cookbook new_resource.solrconfig_cookbook
     owner    new_resource.solr_user
@@ -98,7 +116,7 @@ action :create do
     only_if  { new_resource.use_custom_solrconfig }
   end
 
-  cookbook_file "#{core_conf_dir}/schema.xml" do
+  cookbook_file "#{core_dir}/conf/schema.xml" do
     source   new_resource.schema_source
     cookbook new_resource.schema_cookbook
     owner    new_resource.solr_user