+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This file is used to list changes made in each version of the app_solr cookbook.
 
+## 2.0.0 - 2022-11-
+### Changed
+- Now supports only Chef 17.x and above
+
+### Added
+- Support for Solr 5.x
+- Option to not set user ulimits
+
+### Removed
+- Non-user-interactive resource properties
+
 ## 1.1.0 - 2021-11-15
 ### Added
 - Can now install Solr 6.6.6